remove PowerDynamicsNREL subpackage
