--- PowerDynamicsNREL/examples/OMIB/OMIB.dyr
@@ -1,2 +1,0 @@
-101 'GENCLS' 1  0.000000  0.000000  /
-102 'GENCLS' 1  3.1480000  2.000000  /
