--- PowerDynamicsNREL/examples/OMIB/OMIB.raw
@@ -1,28 +1,0 @@
-0,   100.00,  32, 0, 1, 60.00     / PSS(R)E 32 RAW created by rawd32  TUE, JUL 21 2020  18:43
-
-
-   101,'BUS 1', 230.0000,3,   1,   1,   1,1.05000,   0.0000
-   102,'BUS 2', 230.0000,2,   1,   1,   1,1.04000,   1.3118
- 0 /End of Bus data, Begin Load data
- 0 /End of Load data, Begin Fixed shunt data
- 0 /End of Fixed shunt data, Begin Generator data
-   101,'1 ',   -50.000,    21.572,   100.000,  -100.000,1.05000,     0,   100.000, 0.00000E+0, 1.00000E-5, 0.00000E+0, 0.00000E+0,1.00000,1,  100.0,   100.000,     0.000,   1,1.0000
-   102,'1 ',    50.000,   -20.228,   100.000,  -100.000,1.04000,     0,   100.000, 0.00000E+0, 2.99500E-1, 0.00000E+0, 0.00000E+0,1.00000,1,  100.0,   100.000,     0.000,   1,1.0000
- 0 /End of Generator data, Begin Branch data
-   101,    102,'1 ', 0.00000E+0, 1.00000E-1,   0.00000,  100.00,  100.00,  100.00,  0.00000,  0.00000,  0.00000,  0.00000,1,1,   0.00,   1,1.0000
-   101,    102,'2 ', 0.00000E+0, 1.00000E-1,   0.00000,  100.00,  100.00,  100.00,  0.00000,  0.00000,  0.00000,  0.00000,1,1,   0.00,   1,1.0000
- 0 /End of Branch data, Begin Transformer data
- 0 /End of Transformer data, Begin Area interchange data
- 0 /End of Area interchange data, Begin Two-terminal dc line data
- 0 /End of Two-terminal dc line data, Begin VSC dc line data
- 0 /End of VSC dc line data, Begin Impedance correction table data
- 0 /End of Impedance correction table data, Begin Multi-terminal dc line data
- 0 /End of Multi-terminal dc line data, Begin Multi-section line data
- 0 /End of Multi-section line data, Begin Zone data
- 0 /End of Zone data, Begin Inter-area transfer data
- 0 /End of Inter-area transfer data, Begin Owner data
- 0 /End of Owner data, Begin FACTS device data
- 0 /End of FACTS device data, Begin Switched shunt data
- 0 /End of Switched shunt data, Begin GNE device data
- 0 /End of GNE device data
-Q
