--- PowerDynamicsNREL/examples/ieee14bus/dyn_data.dyr
@@ -1,5 +1,0 @@
-1 'GENCLS' 1  3.1480000  2.000000  /
-2 'GENCLS' 1  3.1480000  2.000000  /
-3 'GENCLS' 1  3.1480000  2.000000  /
-6 'GENCLS' 1  3.1480000  2.000000  /
-8 'GENCLS' 1  3.1480000  2.000000  /
