--- PowerDynamicsNREL/src/PowerDynamicsNREL.jl
@@ -1,37 +1,0 @@
-module PowerDynamicsNREL
-
-using Reexport: @reexport
-
-@reexport using PowerDynamics
-using PowerDynamics.IOComponents
-using PowerSystems
-using PowerSimulationsDynamics
-
-using BlockSystems
-using ModelingToolkit
-using ModelingToolkit: getname, rename, renamespace
-using ModelingToolkit.SymbolicUtils: Symbolic
-using NetworkDynamics
-
-using OrderedCollections: OrderedDict
-
-include("utils.jl")
-
-# include files related to the MetaGenerator model
-include("MetaGenerator/Movers.jl")
-include("MetaGenerator/Shafts.jl")
-include("MetaGenerator/Machines.jl")
-include("MetaGenerator/AVRs.jl")
-include("MetaGenerator/PSSs.jl")
-include("MetaGenerator/MetaGenerator.jl")
-
-# include bus model
-include("Bus.jl")
-
-# include mapping for branch types
-include("Branches.jl")
-
-# include grid constructor
-include("GridConstructor.jl")
-
-end # module
